requirements.txt
requests
PyMuPDF
re2
langchain
langchain-pinecone
langchain-openai
selenium
beautifulsoup4
python-dotenv
langchain_community
pdfminer.six
